Simplify ThreadBaseExceptionFilter
It is always called with `swallowing = TRUE`, so remove dead code.
It is only called from `ThreadBaseExceptionSwallowingFilter` via
`ThreadBaseExceptionAppDomainFilter`, so merge these three functions
together.

Instead of using the `PVOID` paramter to pass a pointer to a
`struct ThreadBaseExceptionFilterParam`, just pass the
`UnhandledExceptionLocation`.

All this cruft seems to be left over from when starting a thread could
entail an AppDomain transition.